Change filter

+++ 407_eth/GP2Y_v3.1_eth/Core/Inc/filter_sma.h
@@ -22,8 +22,8 @@ class filter{
   private:
    uint16_t Filter_Buffer[FILTER_SMA_ORDER] = {0,};
    uint16_t Result = 0;
-   uint16_t gorge = 0;                  // минимальное значение
-   uint16_t peak = 0;                   // пиковое значение
+   float gorge = 0;                  // минимальное значение
+   float peak = 0;                   // пиковое значение
    
    friend class sensor;
